+++ examples/densities/main.rs
@@ -4,10 +4,11 @@ fn main() -> anyhow::Result<()> {
     let values_1: Vec<u32> = (0..200).chain(0..50).collect();
     let values_2: Vec<u32> = (100..300).chain(100..220).chain(150..250).collect();
 
-    pre::Density::new(values_1)
+    let density_1 = pre::Density::new(values_1)
         .set_title("My legend")
-        .to_owned()
-        .compare_with(vec![pre::Density::new(values_2)])
+        .to_owned();
+
+    (density_1 + pre::Density::new(values_2))
         .set_title("My main title")
         .plot("identifier")?;