+++ src/sequence/comparison.rs
@@ -16,6 +16,7 @@ use crate::errors::PreexplorerError;
 // Traits
 pub use crate::traits::{Configurable, Plotable, Saveable};
 use core::fmt::Display;
+use core::ops::{Add, AddAssign};
 
 /// Comparison counter part of ``Sequence`` struct.
 ///
@@ -52,15 +53,48 @@ where
     }
 }
 
-impl<T> crate::traits::Comparison<crate::sequence::Sequence<T>> for Sequences<T>
+impl<T> Add<crate::Sequence<T>> for Sequences<T>  
 where
     T: Display + Clone,
 {
-    fn add(&mut self, other: crate::sequence::Sequence<T>) -> &mut Self {
-        self.data_set.push(other);
+    type Output = Self;
+
+    fn add(mut self, other: crate::Sequence<T>) -> Self { 
+        self += other;
         self
     }
 }
+
+impl<T> Add for Sequences<T>  
+where
+    T: Display + Clone,
+{
+    type Output = Self;
+
+    fn add(mut self, other: Self) -> Self { 
+        self += other;
+        self
+    }
+}
+
+impl<T> AddAssign<crate::Sequence<T>> for Sequences<T>  
+where
+    T: Display + Clone,
+{
+    fn add_assign(&mut self, other: crate::Sequence<T>) { 
+        self.data_set.push(other);
+    }
+}
+
+impl<T> AddAssign for Sequences<T>  
+where
+    T: Display + Clone,
+{
+    fn add_assign(&mut self, mut other: Self) { 
+        self.data_set.append(&mut other.data_set);
+    }
+}
+
 
 impl<T> Configurable for Sequences<T>
 where